added testing for menu and booking

+++ restaurant/models.py
@@ -14,7 +14,7 @@ class Booking(models.Model):
 
 class Menu(models.Model):
    name = models.CharField(max_length=200) 
-   price = models.IntegerField(null=False) 
+   price = models.DecimalField(max_digits=10, decimal_places=2, null=False) 
    menu_item_description = models.TextField(max_length=1000, default='') 
    inventory = models.IntegerField(default=0)
